FSA-5722: Adds page check and move function to another file
