+++ projects/fsastorefront-e2e-cypress/cypress/helpers/checkout/checkout-steps.ts
@@ -273,3 +273,24 @@ export function checkSyncPilotComparisonTable() {
       cy.get('p.mb-0').should('contain.text', 'Speak to an Agent');
     });
 }
+
+export var categoryPage = {
+  homeowners: 'insurance_main_homeowners',
+  renters: 'insurance_main_renters',
+  auto: 'insurance_main_auto',
+  life: 'insurance_main_life',
+  travel: 'insurance_main_travel',
+  event: 'insurance_main_event',
+  savings: 'insurance_main_savings',
+  currentAccount: 'banking_main_current_account',
+  creditCard: 'banking_main_credit_card',
+  loan: 'banking_main_loans',
+  ftd: 'banking_main_fixed_term_deposits',
+  userIdentification: 'user-identification',
+};
+
+export function checkPageURL(page: string) {
+  cy.location().should(loc => {
+    expect(loc.href).to.include(`${page}`);
+  });
+}
